LayoutTests/fast/mediastream/enumerateDevices-camera-denied-expose-with-capture.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
promise_test(async (test) => {
    const stream = await navigator.mediaDevices.getUserMedia({ audio : true });
    test.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));

    if (window.testRunner)
       testRunner.resetUserMediaPermission();
    const devices = await navigator.mediaDevices.enumerateDevices();
    for (const device of devices)
        assert_true(device.label === "" || device.kind !== "videoinput");
}, "enumerateDevices() filters out cameras if camera permission is prompt");

promise_test(async (test) => {
    if (window.testRunner)
        testRunner.setMicrophonePermission(false);

    await navigator.mediaDevices.getUserMedia({ audio : true });
    const devices = await navigator.mediaDevices.enumerateDevices();
    for (const device of devices)
        assert_true(device.label === "" || device.kind !== "videoinput");
}, "enumerateDevices() filters out cameras if camera permission is denied");

promise_test(async (test) => {
    if (window.testRunner)
        testRunner.setMicrophonePermission(true);

    await navigator.mediaDevices.getUserMedia({ audio : true });
    const devices = await navigator.mediaDevices.enumerateDevices();
    for (const device of devices)
        assert_true(device.label === "" || device.kind !== "videoinput");
}, "enumerateDevices() filters out devices even if camera permission is granted and microphone is in use");

    </script>
</head>
<body>
</body>
</html>
