LayoutTests/fast/mediastream/enumerateDevices-camera-denied-expose-with-capture-expected.txt


PASS enumerateDevices() filters out cameras if camera permission is prompt
PASS enumerateDevices() filters out cameras if camera permission is denied
PASS enumerateDevices() filters out devices even if camera permission is granted and microphone is in use
